Refactor: Backend Ticket Master Caching
Event parsing logic has been moved to the backend as the caching process demanded that only parsed events be stored in our database. As a result, many things broke and had to be put back together again. The TM api logic was moved to a helper function along side the tm response parsing logic, the artist event processing logic, the caching retrieval logic and certainly some others.

At this point it appears that the system is correctly fetching artist data from ticket master and storing results in the database for 1 month. If there are no results for an artist, the system will trigger 1 automatic re-attempt.

Additionally, api fetching logic has been /mostly/ converted to be array-size agnostic; it should theoretically work when there are more than two artists submitted for an overlap. This is stage one for allowing users to specify a series of artists and see overlaps that exist across any of them.

NOTE: There is no spotify image caching yet.